clean up metadata
