+++ app/src/app/(interactive)/map/edit/[map_id]/page.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
 import MapPage from '@/app/components/MapPage/MapPage';
 import { generateMapPageMetadata } from '@/app/utils/metadata/pageMetadataUtils';
-import { MapPageProps } from '../types';
+import { MapPageProps } from '../../types';
+
 
 export const generateMetadata = generateMapPageMetadata
 
-export default function ViewPage({params}: MapPageProps) {
+export default function EditPage({params}: MapPageProps) {
   const {map_id} = params;
-  return <MapPage isEditing={false} mapId={map_id} />;
+  return <MapPage isEditing={true} mapId={map_id} />;
 }
