+++ app/src/app/(interactive)/map/types.ts
@@ -1,0 +1,5 @@
+export interface MapPageProps {
+  params: {
+    map_id: string;
+  };
+}
